Small updates for tests
